Create Pages, install font awesome library Delete pages folders
