+++ pages/about.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const about = () => {
+  return (
+    <div>about</div>
+  )
+}
+
+export default about
