+++ pages/blog.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const blog = () => {
+  return (
+    <div>blog</div>
+  )
+}
+
+export default blog
