+++ pages/contact.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const contact = () => {
+  return (
+    <div>contact</div>
+  )
+}
+
+export default contact
