+++ src/Components/NavBar/NavBar.jsx
@@ -1,9 +1,12 @@
+import { faThumbsUp } from '@fortawesome/free-solid-svg-icons'
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import React from 'react'
 
 const NavBar = () => {
   return (
     <div>
-      
+      Navbar
+      <FontAwesomeIcon style={{width:'30px'}} icon={faThumbsUp} />
     </div>
   )
 }
