--- src/app/about/page.jsx
@@ -1,11 +1,0 @@
-
-
-const about = () => {
-  return (
-    <div>
-      
-    </div>
-  )
-}
-
-export default about
