--- src/app/blog/page.jsx
@@ -1,11 +1,0 @@
-
-
-const blog = () => {
-  return (
-    <div>
-      
-    </div>
-  )
-}
-
-export default blog
