--- src/app/contact/page.jsx
@@ -1,11 +1,0 @@
-
-
-const contact = () => {
-  return (
-    <div>
-      
-    </div>
-  )
-}
-
-export default contact
